feat(widget-confirmation): add switching assistant section
