+++ apps/store/src/components/ConfirmationPage/SwitchingAssistantSection/SwitchingAssistantSection.types.ts
@@ -1,0 +1,4 @@
+export type SwitchingAssistantData = {
+  shopSessionOutcomeId: string
+  companyDisplayName: string
+}
